+++ onsenui/esm/elements/ons-alert-dialog.js
@@ -142,6 +142,13 @@ export default class AlertDialogElement extends BaseDialogElement {
    */
 
   /**
+   * @event dialogcancel
+   * @description
+   * [en]Fired when the dialog is canceled.[/en]
+   * [ja][/ja]
+   */
+
+  /**
    * @attribute modifier
    * @type {String}
    * @description