+++ onsenui/esm/elements/ons-dialog.spec.js
@@ -92,9 +92,9 @@ describe('OnsDialogElement', () => {
       expect(spy).to.have.been.called.once;
     });
 
-    it('emits a \'dialog-cancel\' event', () => {
+    it('emits a \'dialogcancel\' event', () => {
       const promise = new Promise((resolve) => {
-        dialog.addEventListener('dialog-cancel', resolve);
+        dialog.addEventListener('dialogcancel', resolve);
       });
 
       dialog.setAttribute('cancelable', '');