+++ onsenui/CHANGELOG.md
@@ -22,6 +22,7 @@ dev
  * ons-button: Allow form submission on submit key keypress. ([#2857](https://github.com/OnsenUI/OnsenUI/issues/2857)).
  * ons-list-item: Clicks are ignored while the expandable list item is expanding/contracting.
  * ons-button, ons-alert-dialog-button, ons-toolbar-button, ons-action-sheet-button: Ensure ons-ripple is imported when button is imported.
+ * ons-dialog, ons-alert-dialog, ons-modal, ons-popover, ons-action-sheet: `dialog-cancel` event is renamed to `dialogcancel`.
 
  ### New Features